+++ src/main/java/de/samply/frontend/Action.java
@@ -1,9 +1,5 @@
 package de.samply.frontend;
 
-public record Action(
-        Site site,
-        Module module,
-        String path
-) {
+public enum Action {
 
 }